Model can be made for a single player.

+++ app/training/keras/data_preparation.py
@@ -22,17 +22,26 @@ def _filter_by_sufficient_no_ratings(df):
     return df.reset_index(drop=True)
 
 
-def _to_trained_on(df, key, ModelClass):
+def _to_trained_on(df, key, ModelClass, rebuild):
     ids = pd.Series(df[key].unique())
-    ModelClass.objects.all().delete()
-    trained_on = []
-    for i, id_ in enumerate(ids):
-        kwds = {'id_in_model': i, key: id_}
-        trained_on.append(ModelClass(**kwds))
-    ModelClass.objects.bulk_create(trained_on)
+    if rebuild:
+        ModelClass.objects.all().delete()
+        trained_on = []
+        for i, id_ in enumerate(ids):
+            kwds = {'id_in_model': i, key: id_}
+            trained_on.append(ModelClass(**kwds))
+        ModelClass.objects.bulk_create(trained_on)
+        id_to_model_indice = pd.Series(ids.index, index=ids)
+    else:
+        kwds = {'{}__in'.format(key): ids}
+        trained_on = ModelClass.objects.all().filter(**kwds)
+        trained_on_df = pd.DataFrame.from_records(trained_on.values())
+        id_to_model_indice = pd.Series(
+            list(trained_on_df.id_in_model), index=trained_on_df[key]
+        )
 
-    id_to_model_indice = pd.Series(ids.index, index=ids)
-    df[key] = df[key].map(id_to_model_indice)
+    df.loc[:, key] = df[key].map(id_to_model_indice)
+    df = df.dropna()
     return df
 
 
@@ -40,11 +49,10 @@ def load_dataset():
     qs = Rating.objects.all()
     df = pd.DataFrame.from_records(qs.values()).dropna()
     df = _filter_by_sufficient_no_ratings(df)
-    np.random.shuffle(df.values)
     return df
 
 
-def to_keras_model_indices(df):
-    df = _to_trained_on(df, 'game_id', TrainedOnGame)
-    df = _to_trained_on(df, 'player_id', TrainedOnPlayer)
+def to_keras_model_indices(df, rebuild_indices=False):
+    df = _to_trained_on(df, 'game_id', TrainedOnGame, rebuild=rebuild_indices)
+    df = _to_trained_on(df, 'player_id', TrainedOnPlayer, rebuild=rebuild_indices)
     return df
